feat: allows selective loading of extensions using optional dependencies
BREAKING CHANGE: The [full] optional dependency needs to be loaded to support the functionality that was previously bundled into the extension. This is to support a minimal install.

Adds a [full] optional dependency in pyproject.toml that is the equivalent of the previous version.
Only sets up extensions that are installed and detected, so other subsets of deps can be supported if necessary.
Changes display_contributors default to False to support the minimal install option.
Changes RTD build to use full package