feat: update copyright statement, add license line (#105)
* footer: update copyright statement, add license line

For #104:

Requires the `license.name` and `license.url` config values.

The format is now the standard: `© <year> <holder>`

Places the license on a separate line and provides an optional link to
it.

Other:

It turns out that `author` isn't part of the standard Sphinx config, so
I copied it into the `html_context`.

* chore: syntax fix for black linting

---------

Co-authored-by: Michael Park <michael.park@canonical.com>